pyproject.toml
[project]
name = "my-college-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.26.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[tool.uv]
link-mode = "copy"
